@@ -1,5 +1,5 @@
 //
-//  Copyright © 2023 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//  Copyright © 2025 Jesús Alfredo Hernández Alarcón. All rights reserved.
 //
 
 import Foundation
@@ -8,12 +8,13 @@ struct WhatsNewState: Equatable {
     var title: String
     var features: [WhatsNewItemViewModel]
 
-    public init(title: String, features: [WhatsNewItemViewModel]) {
+    init(title: String, features: [WhatsNewItemViewModel]) {
         self.title = title
         self.features = features
     }
 }
 
+@MainActor
 public func WhatsNewView(title: String, features: FeaturesBuilder, didTapContinueButton: @escaping () -> Void) -> WhatsNew {
     WhatsNew(
         viewModel: WhatsNewViewModel(
